feat: title alert box and edit style

+++ components/Account.tsx
@@ -37,38 +37,47 @@ const Account = ({ account }: Props) => {
     }
 
     const onDelete = async () => {
-        await fetch(`${process.env.NEXT_PUBLIC_API_URL}/accounts/${id}`, { method: "DELETE" })
-        router.refresh()
+        if (window.confirm(`Êtes vous sûr de vouloir supprimer ${account.name} ?`)) {
+            await fetch(`${process.env.NEXT_PUBLIC_API_URL}/accounts/${id}`, { method: "DELETE" })
+            router.refresh()
+        }
     }
 
     return (
-        <div className='flex justify-between items-center text-white font-satoshi'>
-            <div className='w-10'><p className='text-xl'>{id}</p></div>
-            <Link href={`accounts/${id}`} className='w-50'>
-                <p className='text-xl from-green-400 via-blue-500 to-purple-500 bg-gradient-to-br bg-clip-text text-white hover:text-transparent transition cursor-pointer'>{name}</p>
-            </Link>
-            <div className='w-20 mx-6'>
-                <label htmlFor="active">Bot</label>
-                <input className='ml-2' name='active' type="checkbox" defaultChecked={active} onChange={onChange} />
-            </div>
-            <div className='flex flex-col text-center'>
-                <label htmlFor="maxLeverage">Levier max</label>
-                <div className="rounded-full from-green-400 via-blue-500 to-purple-500 bg-gradient-to-br mx-4 p-1 w-30">
-                    <input name='maxLeverage' className='rounded-full p-2 text-black w-full' type="number" value={leverage} onChange={(e) => setLeverage(Number(e.target.value))} />
+        <div>
+            <div className='flex justify-between items-center text-sm lg:text-lg text-white font-satoshi'>
+                <div className='max-w-6'><p className='text-md lg:text-xl'>{id}</p></div>
+                <Link href={`accounts/${id}`} className='w-24'>
+                    <p className='text-md lg:text-xl from-green-400 via-blue-500 to-purple-500 bg-gradient-to-br bg-clip-text text-white hover:text-transparent transition cursor-pointer'>{name}</p>
+                </Link>
+                <div className='max-w-12 mx-6'>
+                    <label htmlFor="active">Bot</label>
+                    <input className='ml-2' name='active' type="checkbox" defaultChecked={active} onChange={onChange} />
+                </div>
+                <div className='flex flex-col text-center w-16 lg:w-24'>
+                    <label htmlFor="maxLeverage">Levier max</label>
+                    <div className="rounded-full from-green-400 via-blue-500 to-purple-500 bg-gradient-to-br mx-2 lg:mx-4 p-1 w-full">
+                        <input name='maxLeverage' className='rounded-full p-1 lg:p-2 text-black w-full' type="number" value={leverage} onChange={(e) => setLeverage(Number(e.target.value))} />
+                    </div>
+                </div>
+                <div className='flex flex-col text-center'>
+                    <label htmlFor="bankrollSize">Taille de position (en %)</label>
+                    <div className="rounded-full from-green-400 via-blue-500 to-purple-500 bg-gradient-to-br mx-2 lg:mx-4 p-1 w-[150px]">
+                        <input name='bankrollSize' className='rounded-full p-1 lg:p-2 text-black w-full' type="number" value={bankroll} onChange={(e) => setBankroll(Number(e.target.value))} />
+                    </div>
+                </div>
+                <div className='mt-5'>
+                    <button className="m-2 lg:m-4 p-1 rounded-full from-green-400 via-blue-500 to-purple-500 bg-gradient-to-br" onClick={onSubmit}>
+                        <span className="block text-black rounded-full px-2 py-1 lg:px-4 lg:py-2 bg-white hover:bg-transparent hover:text-white transition">Valider</span>
+                    </button>
+                </div>
+                <div className='mt-5'>
+                    <button className="m-2 lg:m-4 p-1 rounded-full from-green-400 via-blue-500 to-purple-500 bg-gradient-to-br" onClick={onDelete}>
+                        <span className="block text-black rounded-full px-2 py-1 lg:px-4 lg:py-2 bg-white hover:bg-transparent hover:text-white transition">Supprimer</span>
+                    </button>
                 </div>
             </div>
-            <div className='flex flex-col text-center'>
-                <label htmlFor="bankrollSize">Taille de position (en %)</label>
-                <div className="rounded-full from-green-400 via-blue-500 to-purple-500 bg-gradient-to-br mx-4 p-1 w-30">
-                    <input name='bankrollSize' className='rounded-full p-2 text-black w-full' type="number" value={bankroll} onChange={(e) => setBankroll(Number(e.target.value))} />
-                </div>
-            </div>
-            <div className='mt-5'><button className="m-4 p-1 rounded-full from-green-400 via-blue-500 to-purple-500 bg-gradient-to-br" onClick={onSubmit}>
-                <span className="block text-black rounded-full px-4 py-2 bg-white hover:bg-transparent hover:text-white transition">Valider</span>
-            </button></div>
-            <div className='mt-5'><button className="m-4 p-1 rounded-full from-green-400 via-blue-500 to-purple-500 bg-gradient-to-br" onClick={onDelete}>
-                <span className="block text-black rounded-full px-4 py-2 bg-white hover:bg-transparent hover:text-white transition">Supprimer</span>
-            </button></div>
+            <div className="from-green-400 via-blue-500 to-purple-500 bg-gradient-to-r h-0.5 my-2" />
         </div>
     )
 }
